+++ metadata.rb
@@ -6,7 +6,7 @@ license          "Apache License, Version 2.0"
 description      "LWRPs to generate services using native system features and service managers. (SysV, Upstart, ...)"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 
-version          "0.1.2"
+version          "0.1.3"
 
 depends          "unix_bin", ">= 0.2.7"
 depends          "resource_masher", ">= 0.10.0"