+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 0.1.3
+
+* Fixed stock service resource initialization and proxying, fixes #2  (thanks @srenatus)
+* Upstart provider now compliant with Debian policy 9.11.1 - init.d compatibility
+* Improved testing for resource notifications and auto-starting services
+* Added support/testing for removal of PID files after services are stopped
+* Further integration of run_action_now cookbook
+* Added some examples to documentation
+
 ## Version 0.1.2
 
 * Fixed SysV services not returning proper status codes